+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v1.1.0
+
+### Features
+
+- Add automatic prerelease tagging for prerelease versions.
+- Enforce prerelease versions when the `--prerelease` option is set.
+
 ## v1.0.0
 
 ### Major Release